Select filename and keep file extension
